feat: add building selection and draw buildings

+++ src/House.h
@@ -13,7 +13,7 @@ protected:
     sf::Vector2f _house_shape_size = sf::Vector2f(50, 50);
     sf::Color _house_shape_color = sf::Color::Green;
     sf::Color _house_outline_color = sf::Color::White;
-    int _house_shape_outline_thickness = 1;
+    float _house_shape_outline_thickness = 1;
 
     float _house_cost = 30;
     float _house_rentability = 2;
